Moved the privateRoutes.jsx and useToken.jsx to the different folders.
